move config to mognodb.go
